requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
click==6.7
decorator==4.0.11
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.5
django-cors-headers==2.0.2
django-debug-toolbar==1.6
django-filter==1.0.1
django-toolbelt==0.0.1
django-user-agents==0.3.0
djangorestframework==3.6.2
geocoder==1.20.1
gunicorn==19.6.0
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
psycopg2==2.7
PyMySQL==0.7.10
pyparsing==2.1.10
PyYAML==3.12
ratelim==0.1.6
requests==2.13.0
six==1.10.0
sqlparse==0.2.3
static3==0.7.0
ua-parser==0.7.3
user-agents==1.1.0
